requirements.txt
cffi==0.9.2
cryptography==0.8.1
docopt==0.6.2
pyasn1==0.1.7
pycparser==2.10
pyOpenSSL==0.14
PySocks==1.5.3
six==1.9.0
